chore: rename dbgen package files and remove small file
